+++ app/src/main/java/com/example/andoirdduan/Database/DataBase.java
@@ -4,29 +4,52 @@ import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
+import android.database.sqlite.SQLiteStatement;
 
-public class DataBase<TABLE_CREATE> extends SQLiteOpenHelper {
+import com.example.andoirdduan.SanPham.SanPham;
+
+import java.util.ArrayList;
+
+public class DataBase extends SQLiteOpenHelper {
+    ArrayList<SanPham> arraySanPham;
     public DataBase(Context context, String name, SQLiteDatabase.CursorFactory factory, int version) {
         super(context, name, factory, version);
     }
 
-        public void QueryData(String sql) {
-            SQLiteDatabase database = getReadableDatabase();
-            database.execSQL(sql);
-        }
+    public void TruyVan(String sql)
+    {
+        SQLiteDatabase db=getWritableDatabase();
+        db.execSQL(sql);
+    }
 
-        public Cursor GetData(String sql) {
-            SQLiteDatabase database = getWritableDatabase();
-            return database.rawQuery(sql, null);
-        }
+    public Cursor TruyVanTraVe(String sql)
+    {
+        SQLiteDatabase db=getWritableDatabase();
+        return db.rawQuery(sql,null);
+    }
 
-        @Override
-        public void onCreate(SQLiteDatabase sqLiteDatabase) {
+    @Override
+    public void onCreate(SQLiteDatabase db) {
 
     }
 
     @Override
-    public void onUpgrade(SQLiteDatabase sqLiteDatabase, int i, int i1) {
+    public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+
+    }
+    public void Insert(String maSP, String tenSP, String phanLoai, int soLuong, int gia, String moTa, byte[] hinh) {
+        SQLiteDatabase db=getWritableDatabase();
+        String sql= "Insert into SanPham values (?,?,?,?,?,?,?)";
+        SQLiteStatement statement=db.compileStatement(sql);
+        statement.clearBindings();
+        statement.bindString(1,maSP);
+        statement.bindString(2,tenSP);
+        statement.bindString(3,phanLoai);
+        statement.bindLong(4,soLuong);
+        statement.bindLong(5,gia);
+        statement.bindString(6,moTa);
+        statement.bindBlob(7,hinh);
+        statement.executeInsert();
     }
 
 }
