+++ app/src/main/java/com/example/andoirdduan/Login/LoadingScreenActivity.java
@@ -7,19 +7,18 @@ import android.widget.Button;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.andoirdduan.Database.DataBase;
 import com.example.andoirdduan.R;
-import com.example.andoirdduan.SanPham.SQL;
+import com.example.andoirdduan.Database.DataBase;
 
 public class LoadingScreenActivity extends AppCompatActivity {
-    public static SQL db;
+    public static DataBase db;
     Button add;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView( R.layout.loading_screen_activity);
 
-        db = new SQL(this,"ShopGiay.sqlite",null,1);
+        db = new DataBase(this,"ShopGiay.sqlite",null,1);
         db.TruyVan("Create Table If not Exists SanPham(ID Text Primary Key, tenSP Text, theLoai Text,soLuong INTEGER, giaTien INTEGER, moTa Text, hinhAnh Blob)");
         add = findViewById(R.id.btnRegister_RegisterActivity);
         add.setOnClickListener(new View.OnClickListener() {
