+++ app/src/main/java/com/example/andoirdduan/Login/LoginActivity.java
@@ -9,7 +9,7 @@ import android.widget.Button;
 import android.widget.TextView;
 
 import com.example.andoirdduan.Database.DataBase;
-import com.example.andoirdduan.HomePage;
+import com.example.andoirdduan.Home.HomePage;
 import com.example.andoirdduan.R;
 
 public class LoginActivity extends AppCompatActivity {
@@ -23,7 +23,7 @@ public class LoginActivity extends AppCompatActivity {
 
         //truy vấn database
         dataBase = new DataBase( LoginActivity.this, "SQLite.sqlite", null, 1 );
-        dataBase.QueryData( "CREATE TABLE IF NOT EXISTS SanPham1(maSP varchar(10), theLoai varchar(20), nhaCungCap varchar(20),tenSP varchar(20), giaTien varchar(20), moTa varchar(30), hinhAnh BLOB)" );
+        dataBase.TruyVan( "CREATE TABLE IF NOT EXISTS SanPham1(maSP varchar(10), theLoai varchar(20), nhaCungCap varchar(20),tenSP varchar(20), giaTien varchar(20), moTa varchar(30), hinhAnh BLOB)" );
 
         DangNhap = findViewById(R.id.btnRegister_RegisterActivity);
         DangKi = findViewById(R.id.tvRegister_LoginActivity);
