+++ app/src/main/java/com/example/andoirdduan/NhanVien/NhanVienActivity.java
@@ -1,0 +1,17 @@
+package com.example.andoirdduan.NhanVien;
+
+import androidx.appcompat.app.AppCompatActivity;
+import android.os.Bundle;
+
+import com.example.andoirdduan.R;
+
+public class NhanVienActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate( savedInstanceState );
+        setContentView( R.layout.activity_nhan_vien );
+
+
+    }
+}
