+++ app/src/main/res/layout/homepage_activity.xml
@@ -3,15 +3,12 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    >
-
+    android:layout_height="match_parent">
     <LinearLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_margin="16dp"
         android:orientation="vertical">
-
         <TextView
             android:id="@+id/textView3"
             android:layout_marginTop="10dp"
@@ -19,7 +16,6 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="Welcome back" />
-
         <TextView
             android:id="@+id/tvUserName"
             android:layout_width="wrap_content"
@@ -29,13 +25,11 @@
             android:layout_marginLeft="10dp"
             android:text="Name"
             android:textSize="28sp" />
-
         <ImageView
             android:id="@+id/imageView3"
             android:layout_width="match_parent"
             android:layout_height="195dp"
             app:srcCompat="@drawable/shoes_slide" />
-
         <TextView
             android:id="@+id/textView7"
             android:layout_width="wrap_content"
@@ -44,7 +38,6 @@
             android:layout_marginLeft="10dp"
             android:textSize="28dp"
             android:textStyle="bold" />
-
         <HorizontalScrollView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content">
@@ -54,7 +47,7 @@
                 android:layout_height="wrap_content"
                 android:orientation="horizontal">
                 <Button
-                    android:id="@+id/C"
+                    android:id="@+id/Size37"
                     android:layout_width="50dp"
                     android:layout_height="50dp"
                     android:backgroundTint="@color/white"
@@ -63,7 +56,6 @@
                     android:textSize="12dp"
                     android:text="item1"
                     android:textColor="@color/black" />
-
                 <Button
                     android:id="@+id/Co"
                     android:layout_width="50dp"
@@ -74,7 +66,6 @@
                     android:textSize="12dp"
                     android:text="item1"
                     android:textColor="@color/black" />
-
                 <Button
                     android:id="@+id/L"
                     android:layout_width="50dp"
@@ -85,7 +76,6 @@
                     android:textSize="12dp"
                     android:text="item1"
                     android:textColor="@color/black"  />
-
                 <Button
                     android:id="@+id/Ha"
                     android:layout_width="50dp"
@@ -96,7 +86,6 @@
                     android:textSize="12dp"
                     android:text="item1"
                     android:textColor="@color/black" />
-
                 <Button
                     android:id="@+id/Hah"
                     android:layout_width="50dp"
@@ -145,9 +134,7 @@
                     android:text="item1"
                     android:textColor="@color/black"/>
             </LinearLayout>
-
         </HorizontalScrollView>
-
         <GridView
             android:id="@+id/gvSanPham"
             android:layout_width="match_parent"
@@ -155,10 +142,7 @@
             android:layout_marginBottom="45dp"
             android:numColumns="2"
             tools:listitem="@layout/cardview_activity"></GridView>
-
-
     </LinearLayout>
-
     <com.google.android.material.bottomappbar.BottomAppBar
         android:id="@+id/bottomMenu"
         android:layout_width="match_parent"
@@ -167,7 +151,6 @@
         app:fabCradleMargin="10dp"
         app:fabCradleRoundedCornerRadius="10dp"
         app:fabCradleVerticalOffset="10dp">
-
         <com.google.android.material.bottomnavigation.BottomNavigationView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -176,17 +159,12 @@
             app:menu="@menu/menu_bottom_nav"
             android:background="@drawable/trans"/>
     </com.google.android.material.bottomappbar.BottomAppBar>
-
     <com.google.android.material.floatingactionbutton.FloatingActionButton
-
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:id="@+id/btnThem_HomePage"
         android:src="@drawable/add_ic"
         app:backgroundTint="#FCD400"
-        app:layout_anchor="@id/bottomMenu"
-        >
-
+        app:layout_anchor="@id/bottomMenu">
     </com.google.android.material.floatingactionbutton.FloatingActionButton>
-
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
