+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,5 @@
 <resources>
-    <string name="app_name">AndoirdDuAn</string>
+    <string name="app_name"></string>
+    <string name="open">Open</string>
+    <string name="close">Close</string>
 </resources>
